computing stabilizer of cubic surfaces using nauty

+++ src/lib/foundations/graph_theory_nauty/nauty.c
@@ -252,7 +252,7 @@ int *aut;
 int aut_counter;
 int *base;
 int base_length;
-int ago;
+long int ago;
 int *transversal_length;
 /* Abdullah 2019 */
